+++ src/app/ApplicationTemplate.Access/NativePlatform/Email/EmailBodyFormat.cs
@@ -1,7 +1,0 @@
-﻿namespace ApplicationTemplate.DataAccess;
-
-public enum EmailBodyFormat
-{
-	PlainText,
-	Html,
-}