+++ src/app/ApplicationTemplate.Access/PlatformServices/Email/EmailBodyFormat.cs
@@ -1,0 +1,7 @@
+﻿namespace ApplicationTemplate.DataAccess.PlatformServices;
+
+public enum EmailBodyFormat
+{
+	PlainText,
+	Html,
+}