+++ src/app/ApplicationTemplate.Access/PlatformServices/Launcher/LaunchFailedException.cs
@@ -1,0 +1,19 @@
+﻿using System;
+
+namespace ApplicationTemplate.DataAccess.PlatformServices;
+
+/// <summary>
+/// Exception used in <see cref="ILauncherService"/> when launch failed.
+/// </summary>
+public sealed class LaunchFailedException : Exception
+{
+	public LaunchFailedException(string message)
+		: base(message)
+	{
+	}
+
+	public LaunchFailedException(string message, Exception innerException)
+		: base(message, innerException)
+	{
+	}
+}