+++ src/app/ApplicationTemplate.Presentation/ViewModels/DadJokes/DadJokesPageViewModel.cs
@@ -5,6 +5,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using ApplicationTemplate.Business;
 using ApplicationTemplate.DataAccess;
+using ApplicationTemplate.DataAccess.PlatformServices;
 using Chinook.DataLoader;
 using Chinook.DynamicMvvm;
 using Chinook.SectionsNavigation;