+++ src/app/ApplicationTemplate.Presentation/ViewModels/ForcedUpdatePageViewModel.cs
@@ -1,4 +1,5 @@
-﻿using Chinook.DynamicMvvm;
+﻿using ApplicationTemplate.DataAccess.PlatformServices;
+using Chinook.DynamicMvvm;
 
 namespace ApplicationTemplate.Presentation;