+++ src/app/ApplicationTemplate.Shared.Views/Configuration/ApiConfiguration.cs
@@ -1,6 +1,6 @@
 ﻿using System.Net.Http;
 using System.Threading.Tasks;
-using ApplicationTemplate.DataAccess;
+using ApplicationTemplate.DataAccess.PlatformServices;
 using MallardMessageHandlers;
 using Microsoft.Extensions.DependencyInjection;