+++ src/app/ApplicationTemplate.Shared.Views/PlatformServices/DeviceInformation/DeviceInformationProvider.cs
@@ -1,0 +1,12 @@
+﻿using Windows.System.Profile;
+
+namespace ApplicationTemplate.DataAccess.PlatformServices;
+
+public sealed class DeviceInformationProvider : IDeviceInformationProvider
+{
+	public string DeviceType => AnalyticsInfo.DeviceForm;
+
+	public string OperatingSystem => AnalyticsInfo.VersionInfo.DeviceFamily;
+
+	public string OperatingSystemVersion => AnalyticsInfo.VersionInfo.DeviceFamilyVersion;
+}