+++ src/app/ApplicationTemplate.Tests.Functional/FunctionalTestBase.cs
@@ -5,6 +5,7 @@ using System.Reactive.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using ApplicationTemplate.DataAccess;
+using ApplicationTemplate.DataAccess.PlatformServices;
 using Chinook.BackButtonManager;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;